Code cleanup & simplification
- Added an interface for raw file operations
- Deleted the two raw file operation files
- Added dependency injection for the raw file service
- Clean up and simplify code
