+++ Call of Duty FastFile Editor/Call of Duty FastFile Editor.csproj
@@ -13,6 +13,7 @@
     <PackageReference Include="DotNetZip" Version="1.16.0" />
     <PackageReference Include="ICSharpCode.TextEditor" Version="3.2.1.6466" />
     <PackageReference Include="ICSharpCode.TextEditorEx" Version="1.3.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.4" />
   </ItemGroup>
 
   <ItemGroup>
