+++ Call of Duty FastFile Editor/Program.cs
@@ -1,3 +1,8 @@
+using System;
+using System.Windows.Forms;
+using Microsoft.Extensions.DependencyInjection;
+using Call_of_Duty_FastFile_Editor.Services;
+
 namespace Call_of_Duty_FastFile_Editor
 {
     internal static class Program
@@ -8,10 +13,22 @@ namespace Call_of_Duty_FastFile_Editor
         [STAThread]
         static void Main()
         {
-            // To customize application configuration such as set high DPI settings or default font,
-            // see https://aka.ms/applicationconfiguration.
             ApplicationConfiguration.Initialize();
-            Application.Run(new MainWindowForm());
+
+            // Create the service collection
+            var services = new ServiceCollection();
+
+            // Register your application services
+            services.AddSingleton<IRawFileService, RawFileService>();
+
+            // Register your forms so they get DependencyInjected
+            services.AddSingleton<MainWindowForm>();
+
+            // Build the provider
+            var provider = services.BuildServiceProvider();
+
+            // 5) Run the app by resolving MainWindowForm from the container
+            Application.Run(provider.GetRequiredService<MainWindowForm>());
         }
     }
 }
